Always used size_t for num_bits and ubits should return an unsigned int.

+++ ais14.cpp
@@ -9,7 +9,7 @@ Ais14::Ais14(const char *nmea_payload, const size_t pad)
 
   assert(message_id == 14);
 
-  const int num_bits = strlen(nmea_payload) * 6 - pad;
+  const size_t num_bits = strlen(nmea_payload) * 6 - pad;
   if (num_bits < 46 || num_bits > 1008) {
     status = AIS_ERR_BAD_BIT_COUNT;
     return;
